Rename hal_402_device_mgr -> hw_device_mgr
This package now does more than just manage drive CiA 402 state
machines in HAL.  It also manages CiA 301 devices with SDOs and PDOs,
EtherCAT devices with object dictionary parsed from ESI files, and
more, and may even be used without HAL.
