+++ scripts/hal_402_mgr
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-from hal_402_device_mgr.hal_402_mgr import Hal402Mgr
+from hw_device_mgr.hal_402_mgr import Hal402Mgr
 
 if __name__ == '__main__':
     hal_402_drives_mgr = Hal402Mgr()
