+++ setup.cfg
@@ -4,4 +4,4 @@ test=pytest
 [tool:pytest]
 # https://pytest-cov.readthedocs.io/en/latest/config.html
 # Uncomment this to run coverage
-# addopts = --cov=hal_402_device_mgr --cov-report term-missing
+# addopts = --cov=hw_device_mgr --cov-report term-missing
