+++ setup.py
@@ -4,7 +4,8 @@ from distutils.core import setup
 from catkin_pkg.python_setup import generate_distutils_setup
 
 d = generate_distutils_setup(
-    packages=['hal_402_device_mgr'], package_dir={'': 'src'}
+    packages=["hw_device_mgr"], package_dir={"": "src"}
+
 )
 
 setup(**d)
